+++ common/kernels/imageBlur_kernel.h
@@ -1,1 +1,1 @@
-__global__ void blurKernel_st1(int* inputPixels, float* intermediate, int* weightedKernel,float* hashes, float* threadReads,uint width, uint height,cudaTextureObject_t tref,int TEX/*, other arguments */);
+__global__ void blurKernel_st1(int* inputPixels, int* intermediate, int* weightedKernel,int* hashes, int* threadReads,uint width, uint height,cudaTextureObject_t tref,int TEX/*, other arguments */);
