+++ common/kernels/imageResize_kernel.cu
@@ -10,7 +10,7 @@
 #define DIAMETER (2*RADIUS + 1)
 #define SINGLEDIMINDEX(i,j,width) ((i)*(width) + (j))
 
-__global__ void resizeKernel_st1(float* inputPixels,float* intermediate, int* weightedKernel,uint height, uint width,uint resizedRows,uint resizedCols,cudaTextureObject_t tref,int NUM_TEX/*, other arguments */)
+__global__ void resizeKernel_st1(int* inputPixels,int* intermediate, int* weightedKernel,uint height, uint width,uint resizedRows,uint resizedCols,cudaTextureObject_t tref,int NUM_TEX/*, other arguments */)
 {
     // assign id's
     int i = blockIdx.x * blockDim.x + threadIdx.x;
@@ -36,7 +36,7 @@ __global__ void resizeKernel_st1(float* inputPixels,float* intermediate, int* we
         j *= yScale;
     }
 
-    float kernelSum = 16.0;
+    int kernelSum = 16.0;
 
     extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
@@ -51,7 +51,7 @@ __global__ void resizeKernel_st1(float* inputPixels,float* intermediate, int* we
             if( ((jdx == j + yScale) && ymod == 0) ||
                 ((jdx == j + yScale) && height <= totalY)) break; // same corner cas
 
-            float tmp = 0.0;
+            int tmp = 0;
             if(jdx < height-2 && jdx >= 2
             && idx < resizedCols-2 && idx >= 0) { // bounds check #1
                 int elemToWrite = SINGLEDIMINDEX(jdx,idx,resizedCols);
@@ -61,7 +61,7 @@ __global__ void resizeKernel_st1(float* inputPixels,float* intermediate, int* we
                     int location = elemToReadFrom + ii;
                     // bounds check #2 for surrounding pix
                     if (location < (width*height) && location >= 0) {
-                        float loaded = inputPixels[location];
+                        int loaded = inputPixels[location];
                         tmp += loaded * weightedKernel[ii];
                         updateGHB(&(ghb[my_ghb_index]),loaded);
                     }
@@ -71,9 +71,9 @@ __global__ void resizeKernel_st1(float* inputPixels,float* intermediate, int* we
                 for(int ii = (DIAMETER-NUM_TEX); ii < DIAMETER;ii++) {
                     int curValueHash = hashGHB(&ghb[my_ghb_index]);
                     int texVal = tex1D<int>(tref,curValueHash);
-                    tmp += (float)(ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
+                    tmp += (int)(ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
                 }
-                float avg = tmp / kernelSum;
+                int avg = tmp / kernelSum;
                 intermediate[elemToWrite] = avg;
             }
         }
